+++ packages/config/src/finder/index.ts
@@ -1,1 +1,0 @@
-export { ConfigFinder } from './config-finder'