+++ packages/config/src/index.ts
@@ -1,5 +1,4 @@
 export * from './bytecode'
 export * from './config'
-export * from './finder'
 export * from './bundler';
 export * from './signature';