+++ examples/my-app/CHANGELOG.md
@@ -1,5 +1,27 @@
 # my-app
 
+## 1.11.0
+
+### Minor Changes
+
+- fix iframe transport
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/provider@1.12.0
+
+## 1.10.0
+
+### Minor Changes
+
+- fix iframe transports
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/provider@1.11.0
+
 ## 1.9.1
 
 ### Patch Changes