+++ packages/0xsodium/CHANGELOG.md
@@ -1,5 +1,45 @@
 # 0xsodium
 
+## 1.11.0
+
+### Minor Changes
+
+- fix iframe transport
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/abi@1.12.0
+  - @0xsodium/auth@1.12.0
+  - @0xsodium/config@1.12.0
+  - @0xsodium/guard@1.12.0
+  - @0xsodium/multicall@1.12.0
+  - @0xsodium/network@1.12.0
+  - @0xsodium/provider@1.12.0
+  - @0xsodium/transactions@1.12.0
+  - @0xsodium/utils@1.12.0
+  - @0xsodium/wallet@1.12.0
+
+## 1.10.0
+
+### Minor Changes
+
+- fix iframe transports
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/auth@1.11.0
+  - @0xsodium/config@1.11.0
+  - @0xsodium/provider@1.11.0
+  - @0xsodium/wallet@1.11.0
+  - @0xsodium/abi@1.11.0
+  - @0xsodium/guard@1.11.0
+  - @0xsodium/multicall@1.11.0
+  - @0xsodium/network@1.11.0
+  - @0xsodium/transactions@1.11.0
+  - @0xsodium/utils@1.11.0
+
 ## 1.9.1
 
 ### Patch Changes